backend/requirements.txt
fastapi
uvicorn
python-multipart
speechrecognition
gTTS
pydub
python-dotenv
requests
pydantic
langchain
faiss-cpu
transformers
sentence-transformers
